Fix compilation error for MacOs test builds
But panic in the implementation instead.

The long-term goal of this change - make it possible to run miniquad based apps
headless without any rendering and any system dependencies.

This is the first step - implementation will panic on runtime and there is no way
to choose dummy backend on linux/windows/wasm. But this allow to run tests on
unsupported platforms.