Textures and buffers delete function
This is very controversial change. Usually rust's graphics libraries provides
RAII objects instead with cleaning up code in Drop. But miniquad is intended to
be used in building type-safe abstractions, not being type-safe by itself.
However miniquad is following Rust's meaning of safety - so no UB on CPU sides
should be possible. Fortunately using deleted textures later is not an UB in
OpenGL.

Maybe for other backends this functions will be marked as unsafe though - it may
be an UB in metal.