Deprecate "miniquad::rand"
While miniquad is still linked to libc on most platforms, libc's rand have a lot
of problems - implementation is different on windows and linux. And non-libc var
iant on JS is not even seedable.

The new recommended way to get random in miniquad - "quad-rand" crate.
quad-rand is exactly the same PRNG algorithm that was in linux's libc, but implem
nted with std atmoics and working in exactly the same way across all target platf
orms.