native/sapp-wasm: Fix undefined error in firefox on version mismatch
When wasm required a function not implemented in importObject
"instantiateStreaming" should throw a LinkError. It works just fine in Chrome and Safari, but not if Firefox.

In firefox it was just an "undefined". This change split  WASM loading into
module compilation and instanciating. This allows to check wich function is
missing in the module and patch importObject with stabs.

Later on this will allow to check gl.js and wasm version match between compile
and instantiate.