native/sapp-windows: Implement cursor grabbing and raw mouse events for grabbed cursor
Cursor grabbing on Windows includes two separate things: 

- limit cursor movement to avoid losing focus on the window
- receiving events from the mouse even if the cursor is stuck on the screen border

The first is done with winapi's ClipCursor. This change guarantees that while the mouse is grabbed cursor will be clipped correctly after window move/resize etc. 

The second is done with RegisterRawInputDevices. RegisterRawInputDevices is used only when the cursor is grabbed is going to re-register each cursor grabbing state is changed. 
So raw device events will be sent only when the cursor is grabbed, which is exactly what we need and match web behavior.