native/sapp-darwin: Fix high-dpi bug
The bug: scale_factor was not updated until first window resize.
It is a blind attempt to fix it. Based on diffing sokol_app with upstream version and on the help from the #169 gh issue.
It looks like there was a bug in sokol_app - _sapp_macos_update_dimensions should be called at the end of "applicationDidFinishLaunching" to set scale factors at the very first app frame