native/sapp-wasm: glReadPixels to gl.js
Usually glReadPixels was not used and probably the glReadPixels was usually just stripped from the build, so wasm linker was happy. But when it actually was used in the rust code - it genereted a JS exception because the function was not implemented.

This PR fix this and the lost glReadPixels is now back.