graphics: Fix broken bind cache on buffer update
restore_buffer_binding should have store_buffer_binding a little upfront,
otherwise "restore" is restoring some weird buffer.

It was more or less fine - usually there were checks on what buffer is
binded and if its a wrong buffer - it was just rebind.

But it was a big bug and resulted into broken state.