native/sapp-wasm: Use relative position for touch events
Previously, touch events were not registered relative to their canvas,
but to the document. This means that touch events would only be
positioned correctly if the canvas was placed in the upper left
corner of the document.

This change applies the same logic to touch events as is currently
applied to mouse events and thereby positions touch events correctly.