graphics: remove statefull set_* calls
Blending, Stencil etc are parts of the pipeline and should be manipulated throught a pipeline. Before we got away with direct global opengl state machine manipulations cause there was no technical reason preventing against this. But with metal and possibly other rendering backends in the future - this kind of APIs would be really painfull to simulate.