linux_wayland, linux_x11: Fixed cursor flickering when show_mouse() is used with a non-default cursor.
This addresses cursor flickering in egui-miniquad-0.16.0.

NOTE: This changes behavior of set_cursor(), it won't show reveal a hidden cursor on X11 and Wayland anymore. But the
new behavior is consistent with Windows.

The problem is that neither Wayland nor X11 backend tracks cusor visibility separately from the icon. Every time the
cursor is shown it is being reset to default cursor. And whenever cursor is changed it is being shown. This is different
from Windows, where SetCursor() is independent from ShowCursor().

To address I track these states explicitly in native::linux_wayland::run() and X11Display.

Test steps:
* Run X11 session
* Run demo example from egui-miniquad-0.16.0
* Try resizing an egui window and observe cursor not flickering anymore.
* Run Wayland session in Plasma 6.5.3
* Modify conf of demo to run on Wayland (default is X11)
* Try resizing an egui window and observe cursor not flickering anymore.